feat: added instructions and ratings

+++ lib/domain/entities/recipe.dart
@@ -5,7 +5,7 @@ class Recipe {
   final List<String> topics;
   final List<String> instructions;
   final String imageUrl;
-  final int rating;
+  final double rating;
   final List<String> ingredients;
   final List<String> quantities;
   final bool isFavourite;
@@ -31,7 +31,7 @@ class Recipe {
     String? name,
     List<String>? instructions,
     String? imageUrl,
-    int? rating,
+    double? rating,
     List<String>? ingredients,
     bool? isFavourite,
   }) {
@@ -66,7 +66,7 @@ class Recipe {
       topics: List<String>.from(map["topics"] ?? []),
       instructions: List<String>.from(map['instructions'] ?? []),
       imageUrl: map['imageUrl'] ?? '',
-      rating: map['rating']?.toInt() ?? 0,
+      rating: map['rating']?.toDouble() ?? 0.0,
       ingredients: List<String>.from(map['ingredients'] ?? []),
       quantities: List<String>.from(map['quantities'] ?? []),
       isFavourite: map['isFavourite'] ?? false,
